+++ src/templates/interfaces/CreateComponent.ts
@@ -1,4 +1,5 @@
 export default interface CreateComponent {
   componentName: string;
   fileName?: string;
+  useIPrefixForInterfaces?: boolean;
 }